+++ cli/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "cli"
+version = "0.1.0"
+authors = ["terkwood <38859656+Terkwood@users.noreply.github.com>"]
+edition = "2018"
+
+[dependencies]
+lindenmayer = { path = "../lindenmayer" }
+serde = "1.0.104"
+serde_derive = "1.0.104"
+serde_json = "1.0.47"
