+++ lindenmayer/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod png;
 pub mod svg;
 
+mod draw;
 mod lsys;
 mod parametric;
