+++ lindenmayer/src/model.rs
@@ -1,2 +1,1 @@
-struct Axiom(char);
-struct Rule
+
